+++ tpe/lib/src/Link_TEST.cc
@@ -19,6 +19,7 @@
 
 #include "Collision.hh"
 #include "Link.hh"
+#include "Model.hh"
 
 using namespace ignition;
 using namespace physics;
@@ -36,6 +37,16 @@ TEST(Link, BasicAPI)
 
   link.SetPose(math::Pose3d(1, 2, 3, 0.1, 0.2, 0.3));
   EXPECT_EQ(math::Pose3d(1, 2, 3, 0.1, 0.2, 0.3), link.GetPose());
+
+  Model model;
+  auto modelPose = math::Pose3d(10, 0, 2, 1, 0, 0);
+  model.SetPose(modelPose);
+  Entity &linkEnt = model.AddLink();
+  ASSERT_NE(nullptr, linkEnt.GetParent());
+
+  linkEnt.SetPose(math::Pose3d(0, 0.2, 0.5, 0, 1, 0));
+  EXPECT_EQ(math::Pose3d(10, -0.312675, 2.43845, 1.23686, 0.471978, 0.918989),
+            linkEnt.GetWorldPose());
 
   Link link2;
   EXPECT_NE(link.GetId(), link2.GetId());